removed a bug and further improved the speed of stat_screen_area. Now THRESHOLD is on the area of faces.

+++ commonType.hh
@@ -1,6 +1,6 @@
 #ifndef __COMMON_TYPE_HH__
 #define __COMMON_TYPE_HH__
-typedef double        Coordinate;
+typedef double       Coordinate;
 typedef unsigned int    Index;
-typedef float          Normalf;
+typedef double         Normalf;
 #endif
